- splitting groupUc
